+++ malachite-float/tests/lib.rs
@@ -181,6 +181,7 @@ pub mod float {
         pub mod cbrt_2;
         pub mod champernowne_constant;
         pub mod copeland_erdos_constant;
+        pub mod dottie_number;
         pub mod e;
         pub mod eulers_constant;
         pub mod gauss_constant;